WIP: only display actions for context-specific rules
the idea is to allow staff to view all rules that will
affect a specific context, but only let them act on
(edit, delete) rules that are specific to a context.

This aims to reduce the complexity of the impact of
these actions, and make it easier for staff to
understand what they are acting on when they do.

Context / futher work suggested

This lays the ground work for ensuring that the 'delete'
function will effectively be a simple delete:
it will get rid of the rules sets for the context
selected, and only for that context.

What fallbacks / defaults /etc. do come into play
once the context-specific rule has been deleted
will still be viewable by selecting the read only
"display all rules that apply mode".

I would consider including a message on delete
to recommend the user goes and views the impact
of their changes.