WIP: feat: display rules for every context possible
Probably needs a refactor (or two) - effectively,
allow staff to use "Display all patron categories and
items types" to see each and every single one of these
listed for the selected context.

Effectively, where default rules apply, we are generating
copies of these rules for each library / patron category /
item type combination which they apply to.

Bug noticed: the context columns in the table will
still display "default" instead of the selected patron
or item category if default rules have been used to populate them

Need to address that.
Also need to ensure that the bold/italics for fallback have
carried over - not certain that it is the fact