Bug 10190: feat: display exhaustive rule list
Add a "Display all patron categories and item types" view where for the selected library, for each patron category and item type combination possible, if no specific rule set is found, then the default rule set(s) that apply display instead.

This is so that administrators have the option to easily see exactly what rules are going to apply to any given context. Administrators can use the "filter" feature to select a specific library (instead of viewing defaults for all), narrow the list down to a patron category and/or and item type.

As previously, fallbacks are denoted by the use of bold italic text.

The original view (showing only defined rules sets) was maintained and is now available as "Display defaults and overrides".