Bug 10190: fix: remove delay overwrite
A bug was found where, on edit, the delay field
in the form was populated by minDelay instead of
the existing value for the given context and trigger.

Update setMinDelay so it does not make changes
to the ruleSetToSubmit object, thus preventing
the overwrite, with the form accurately displaying
the delay as assigned.

Also ensures that delay datatype is consistent
(string|null)